pyproject.toml
[tool.poetry]
name = "spms"
version = "0.1.0"
description = ""
authors = ["idabblewith <109533655+idabblewith@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
black = "^23.3.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
django-environ = "^0.10.0"
pandas = "^2.0.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.14.0"
dj-database-url = "^1.3.0"
pyjwt = "^2.6.0"
requests = "^2.28.2"
reportlab = "^4.0.0"
dbca-utils = "^1.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
